move booleanLiteral to right folder
