+++ packages/@n8n/api-types/src/schemas/booleanLiteral.ts
